fix(t6): address PR review — honour sweep, float hotspot, clamp render, doc sync
Verified each review finding; applied the still-valid ones (minimal):
- agent.py: continuous klinotaxis sweep honours the configured
  `sweep_amplitude_mm` (was `max(1.0, ...)`, which silently ignored sub-1.0
  values like the 0.5 default); spec wording + docstring updated to match.
- continuous_2d.py `_generate_food_candidate`: hotspot branch now returns float
  coords, matching the override's `tuple[float, float]`.
- env.py full-grid render: snap-and-clamp the goal cell before indexing (matches
  the already-guarded viewport render site; defends the indexing op).
- pygame_renderer.py (both food sites): snap to a cell first, then view-test and
  place that same cell so edge sources stay consistent.
- temperature.py / oxygen.py / pheromone.py: docstring `position` types synced to
  the widened `Position` / `tuple[float, float]` signatures.

Skipped (not still-valid, brief reasons):
- multi_agent food_history clamp: it's a position *record*, not an array index,
  and continuous sources are placed within [0, grid_size-1] so rounding can't go
  out of bounds — no crash path.
- ForagingParams config validation: already enforced at config load by the
  pydantic `ForagingConfig` (`Literal` mode + `Field(gt=0.0)` on D/assay_time);
  the `D=None → decay_constant` path in `fick_length` is the intended documented
  fallback, so raising on it would be wrong.
- test exact-`==` → `pytest.approx`: the equality is the deliberate
  byte-stability assertion (both sides run identical IEEE ops; deterministic).
- test `-> None` annotations: not enforced for tests (existing tests in the file
  omit them; pre-commit passes) — adding them would be inconsistent.

276 affected tests pass; ruff/pyright/markdownlint green; openspec validate clean.

+++ packages/quantum-nematode/quantumnematode/env/pheromone.py
@@ -131,7 +131,7 @@ class PheromoneField:
 
         Parameters
         ----------
-        position : tuple[int, int]
+        position : tuple[float, float]
             Grid position to query.
         current_step : int
             Current simulation step (for temporal decay calculation).
@@ -166,7 +166,7 @@ class PheromoneField:
 
         Parameters
         ----------
-        position : tuple[int, int]
+        position : tuple[float, float]
             Grid position to query.
         current_step : int
             Current simulation step.
@@ -192,7 +192,7 @@ class PheromoneField:
 
         Parameters
         ----------
-        position : tuple[int, int]
+        position : tuple[float, float]
             Grid position to query.
         current_step : int
             Current simulation step.
